add repository

+++ src/main/java/com/bteplus/draftbus/entity/InputData.java
@@ -18,17 +18,17 @@ public class InputData {
     private Integer calc_year;
     private Integer country_id;
     private Integer city_id;
-    private double discount_rate;
-    private double social_discount_rate;
-    private double inflation_rate;
-    private double exchange_rate;
+    private Double discount_rate;
+    private Double social_discount_rate;
+    private Double inflation_rate;
+    private Double exchange_rate;
     private Integer temperature;
     private Integer humidity;
     private Integer slope;
-    private double charger_construction;
-    private double chargers_number;
-    private double procurement_cost;
-    private double operational_cost;
-    private double maintenance_cost;
+    private Double charger_construction;
+    private Integer chargers_number;
+    private Double procurement_cost;
+    private Double operational_cost;
+    private Double maintenance_cost;
 
 }
